경험치판매 보너스 단위 수정 및 attackinfo 표시, 경험치 배율 998레벨까지 확장 자이언트 3단계 직업 계열 추가 (자이언트/용병/기사)
* feat: 경험치판매 보너스 단위 수정 및 attackinfo 표시, 경험치 배율 998레벨까지 확장

* feat: 자이언트 3단계 직업 계열 추가 (자이언트/용병/기사)

---------

+++ src/main/java/my/prac/core/game/dto/UserBattleContext.java
@@ -91,8 +91,8 @@ public class UserBattleContext {
 	public int expSellHp;
 	public int expSellAtkMin;
 	public int expSellAtkMax;
-	public int expSellCrit;
-	public int expSellCritDmg;
+	public double expSellCrit;
+	public double expSellCritDmg;
     
 	// [OPT-HUNTER] attackInfo 에서 미리 조회한 dropsRows 공유용 (applyDropBonusToContext 중복 조회 방지)
 	public List<HashMap<String,Object>> preDropRows = null;
